+++ app/src/main/java/hu/bme/aut/fitary/data/DomainExercise.kt
@@ -1,6 +1,7 @@
 package hu.bme.aut.fitary.data
 
 data class DomainExercise(
+    val id: Long,
     var name: String = "",
     var reps: Int = 0
 ) {
