+++ app/src/main/java/hu/bme/aut/fitary/data/DomainUser.kt
@@ -1,6 +1,6 @@
-package hu.bme.aut.fitary.dataSource.model
+package hu.bme.aut.fitary.data
 
-data class UserProfile(
+data class DomainUser(
     val id: String,
     val userMail: String,
     val username: String
