+++ app/src/main/java/hu/bme/aut/fitary/data/DomainWorkout.kt
@@ -1,9 +1,9 @@
 package hu.bme.aut.fitary.data
 
-data class Workout (
-    val uid: String? = null,
-    var userName: String? = null,
+data class DomainWorkout (
+    val uid: String,
+    var userName: String,
     var domainExercises: MutableList<DomainExercise> = mutableListOf(),
-    var score: Double? = 0.0,
+    var score: Double = 0.0,
     var comment: String? = null
 )
