+++ app/src/main/java/hu/bme/aut/fitary/dataSource/model/Workout.kt
@@ -1,8 +1,8 @@
 package hu.bme.aut.fitary.dataSource.model
 
 data class Workout (
-    val uid: String? = null,
+    val uid: String,
     val exercisesAndReps: MutableList<Pair<Long, Int>> = mutableListOf(),
-    var comment: String? = null,
-    var score: Double = 0.0
+    var score: Double = 0.0,
+    var comment: String? = null
 )
