+++ app/src/main/java/hu/bme/aut/fitary/ui/createWorkout/CreateWorkoutActivity.kt
@@ -6,7 +6,7 @@ import hu.bme.aut.fitary.BaseActivity
 import hu.bme.aut.fitary.R
 import hu.bme.aut.fitary.adapter.ExerciseAdapter
 import hu.bme.aut.fitary.data.DomainExercise
-import hu.bme.aut.fitary.data.Workout
+import hu.bme.aut.fitary.dataSource.model.Workout
 import hu.bme.aut.fitary.extensions.validateNonEmpty
 import hu.bme.aut.fitary.ui.exerciseDialog.ExerciseDialog
 import kotlinx.android.synthetic.main.activity_create_workout.*
