+++ app/src/main/java/hu/bme/aut/fitary/ui/userWorkouts/UserWorkoutsFragment.kt
@@ -13,7 +13,7 @@ import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.FirebaseDatabase
 import hu.bme.aut.fitary.R
 import hu.bme.aut.fitary.adapter.WorkoutAdapter
-import hu.bme.aut.fitary.data.Workout
+import hu.bme.aut.fitary.data.DomainWorkout
 import kotlinx.android.synthetic.main.fragment_workouts_user.view.*
 
 class UserWorkoutsFragment : Fragment() {
@@ -47,7 +47,7 @@ class UserWorkoutsFragment : Fragment() {
             .getReference("workouts")
             .addChildEventListener(object : ChildEventListener {
                 override fun onChildAdded(dataSnapshot: DataSnapshot, s: String?) {
-                    val newWorkout = dataSnapshot.getValue<Workout>(Workout::class.java)
+                    val newWorkout = dataSnapshot.getValue<DomainWorkout>(DomainWorkout::class.java)
 
                     if (newWorkout?.uid.equals(userId))
                         workoutAdapter.addWorkout(newWorkout)
